Support the 'beforeExit' event in the master process
So that the master can do cleanup before exiting.

The 'beforeExit' event is a useful way of exposing this both because it requires
no additional configuration of `throng` and because the master process will
automatically be kept alive if the user makes asynchronous calls from within
the 'beforeExit' event.

In order for the 'beforeExit' event to be emitted when it is necessary to
forceably disconnect the worker processes, we must not explicitly terminate
the process i.e. by calling `process.exit`. But this is not necessary since
disconnecting the workers will cause the master process to exit anyway.

The cluster’s shutdown signal is exposed for the 'beforeExit' handler to
examine if desired.